+++ lib/MIP/Recipes/Analysis/Cadd.pm
@@ -26,7 +26,7 @@ BEGIN {
     use base qw{ Exporter };
 
     # Set the version for version checking
-    our $VERSION = 1.08;
+    our $VERSION = 1.09;
 
     # Functions and variables which can be optionally exported
     our @EXPORT_OK = qw{ analysis_cadd analysis_cadd_panel };
@@ -123,7 +123,7 @@ sub analysis_cadd {
 
     use MIP::Get::File qw{ get_io_files };
     use MIP::Get::Parameter qw{ get_recipe_attributes get_recipe_resources };
-    use MIP::Program::Gnu::Bash qw{ gnu_export gnu_unset };
+    use MIP::Program::Gnu::Bash qw{ gnu_cd gnu_export gnu_unset };
     use MIP::Program::Gnu::Coreutils qw{ gnu_mkdir };
     use MIP::Parse::File qw{ parse_io_outfiles };
     use MIP::Program::Bcftools qw{ bcftools_annotate bcftools_concat bcftools_view };
@@ -322,6 +322,16 @@ sub analysis_cadd {
                 outfile_path    => $view_outfile_path,
                 output_type     => q{v},
                 stderrfile_path => $stderrfile_path,
+            }
+        );
+        print {$xargsfilehandle} $SEMICOLON . $SPACE;
+
+        ## Attempt to escape snakemake lock
+        gnu_cd(
+            {
+                filehandle             => $xargsfilehandle,
+                directory_path         => $temp_outdir_path{$contig},
+                stderrfile_path_append => $stderrfile_path,
             }
         );
         print {$xargsfilehandle} $SEMICOLON . $SPACE;